requirements.txt
requests
pyautogui
pygetwindow
pywin32
telebot
python-dotenv
matplotlib
